Fix to removing non-date headers with timestamps
Most everything that gets posted to chat has a timestamp data attribute. We'll
need to add a class to uniquely identify a list-header as being a date header.
Additionally, the header's timestamp attribute does not actually hold the date
that it represents. Instead we just parse the contents as a date object.