Fix to up/down arrow keys in multiline messages
Previously we prevented the default event to stop the cursor from moving when we
cycle through messages, but that unintendedly prevents us from scrolling around
after pasting in a multi-line message. We can return whether a cycle took place
to know whether we should stop the key from moving the cursor.