Minor fix to click-to-message functionality
When the .trim() was added to the message, I did not consider that a space is
added initially when you click a name. The only functionality this fixes is if
you click a name and then try to click another without typing anything. While it
isn't a big deal, it's also not a big fix.