Fix scrollToBottom when a horizontal scrollbar is present
Because the scrollHeight - height calculations do not give the actual max scroll
value, we have reverted to using just scrollHeight, which is an obvious
overshoot. That said, the only reason it was changed originally was so that the
scroll fix for chrome could offset it by 1 and then back again, but this can be
worked around by setting it to 0. Because the browser won't repaint immediately,
no stutter should be visible.