Fix webkit scrolling bug again in newer versions
In newer versions of Chrome and Safari, particularly on OSX, the scrolling bug
returns and seems to be impervious to the previous hack. In this case, it
ignores the scrollTop when it is 0 despite the fact that the scrollbar is at the
bottom. To get around this, we just jiggle the scroll height by 1.