Bug fix for topic announcements in rooms that are not active
The event was originally tied to this.activeRoom since it was only sent to the
initiator who obviously had it focused. After the announcement change, if you
had set the room topic in client A, and client B was in that room but had a
different room active, he would have seen a message that the topic was changed,
but in his current room. Now we simply send down the room name so that we can
add the message to that channel.